requirements.txt
# --- Core app ---
streamlit>=1.36
chromadb>=0.5.0
numpy>=1.24
scikit-learn>=1.3
scipy>=1.11

# --- PDF / DOCX handling ---
PyMuPDF>=1.23   # alias: fitz
PyPDF2>=3.0
python-docx>=1.0
Pillow>=10.0
pytesseract>=0.3

# --- OpenAI / LLM ---
openai>=1.30
tiktoken>=0.7

# --- Présentation / Export ---
python-pptx>=0.6

# --- Divers utiles ---
orjson>=3.9
